refactored client-side and some server js based off of this http://spiritconsulting.com.ar/fedex/2010/11/events-with-jquery-nodejs-and-socket-io/
that is an excellent article about how to create custom jquery socket-io events and handle them

the events are now completely jquery evented on the client

also using a much more recently updated and featureful jquery cookie plugin.
Its now COOKIES and not COOKIE. Was other way around at first which brought in a fun (to find) error :/

it now :
- sets cookies based off of server fired events over socket.io
- calls the "call in progress" animation based off of server fired events as well. you can even change the message from the server!

Lastly, started working on the socket.io code for the /call route of the server. It may be broken on this commit